refactored state to context
