+++ client/src/components/ListTodos.js
@@ -1,6 +1,6 @@
-import React, { Fragment, useEffect, useState } from 'react'
+import React, { Fragment, useContext } from 'react'
 import EditTodo from './EditTodo'
-
+import { AppContext } from '../App'
 //TODO Adding and editing a todo updates db, but app is not reflecting those changes
 // the changes are reflected on refresh
 //find a way to send state back to this component.
@@ -8,39 +8,23 @@ import EditTodo from './EditTodo'
 //lift state up to app level, and use useContext
 // useContext hook :)
 const ListTodos = () => {
-  const [todos, setTodos] = useState([])
-
-  const getTodos = async () => {
-    try {
-      const response = await fetch('http://localhost:5000/todos')
-      const jsonData = await response.json()
-      setTodos(jsonData)
-    } catch (err) {
-      console.error(err.message)
-    }
-  }
-
+  const { todos, handleTodoUpdate: updateParent } = useContext(AppContext)
+  
   //delete todo
   const handleDelete = async (id) => {
     try {
       await fetch(`http://localhost:5000/todos/${id}`, {
         method: 'DELETE',
       })
-      setTodos(todos.filter(todo => todo.todo_id !== id))
+      updateParent()
     } catch (err) {
       console.error(err.message)
     }
   }
   
-  useEffect(() => {
-    getTodos()
-  }, [])
-
-  console.log(todos)
-
   return (
     <Fragment> 
-      <table class="table table-striped mt-3">
+      <table className="table table-striped mt-3">
         <thead>
           <tr>
             <th scope="col">id</th>
