+++ client/src/index.js
@@ -1,13 +1,7 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import './index.css'
 import App from './App'
+import { createRoot } from 'react-dom/client'
+const container = document.getElementById('root')
+const root = createRoot(container);
+root.render(<App tab="home" />);
 
-
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-)
